chore: move evm configurator to evm pkg
Simplifies the package structure by moving the evm configurator to the
evm package.
